refactor(lib): channels MarketChannels doesn't make sense anymore, replace with Vec<MarketChannel>

+++ db/src/storage/rocksdb.rs
@@ -348,7 +348,7 @@ mod test {
     fn serde_get_ranged_cf() {
         let table = "rows";
         let db = db(vec![table.to_string()]);
-        let size = 10_i32.pow(3) as i32;
+        let size = 10_i32.pow(3);
         let before = Utc::now();
         let mut items = vec![];
         for i in 0..size {
@@ -370,7 +370,7 @@ mod test {
     fn get_ranged_cf() {
         let table = "rows";
         let db = db(vec![table.to_string()]);
-        let size = 10_i32.pow(3) as i32;
+        let size = 10_i32.pow(3);
         let before = Utc::now();
         let mut items = vec![];
         for i in 0..size {
@@ -400,7 +400,7 @@ mod test {
         init();
         let table = "rows";
         let db = db(vec![table.to_string()]);
-        let size = 10_i32.pow(4) as i32;
+        let size = 10_i32.pow(4);
         let before = Utc::now();
         let mut then = Utc::now();
         let mut items = vec![];
